requirements.txt
numpy
tensorflow
pyopengl